Update extension.ts
VS code complains about the command not existing. I took a look at an example from the VSCode docs and it has a different way of registering the command, so I changed it to be like the example. Not sure if this will actually work, but I don't really know how to test out an unpublished extension, nor do I want to take the time right now to figure it out 😄